Do not use keys to represent principals.
Up till now, resources such as keys and nonces had to carry which encryption
keys could be used to declassify them to public.  This required a complicated
invariant for resources to be well-formed: a principal, represented as a key,
had to ensure that all of this keys readers were not public.

I decided to replace this by a simpler public/secret distinction.  The hope is
that this will make it easier to verify protocols without forcing all
participants to be honest.